refactor(api): convert .js file to .ts file
